requirements.txt
requests
beautifulsoup4
openai
python-dotenv
langchain
langchain_openai
langchainhub
fastapi
uvicorn
joblib
streamlit
colorama
pydantic
transformers
torch
sentencepiece
sentence_transformers
faiss-cpu